# Documentation/dev-guide/apispec/swagger/rpc.swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "swagger": "2.0",
  "info": {
    "title": "etcdserver/etcdserverpb/rpc.proto",
    "version": "version not set"
  },
  "paths": {
...
    "/v3/kv/compaction": {
      "post": {
        "tags": [
          "KV"
        ],
        "summary": "Compact compacts the event history in the etcd key-value store. The key-value\nstore should be periodically compacted or the event history will continue to grow\nindefinitely.",
        "operationId": "Compact",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/etcdserverpbCompactionRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/etcdserverpbCompactionResponse"
            }
          },
          "default": {
            "description": "An unexpected error response",
            "schema": {
              "$ref": "#/definitions/runtimeError"
            }
          }
        }
      }
    },
    "/v3/kv/deleterange": {
      "post": {
        "tags": [
          "KV"
        ],
        "summary": "DeleteRange deletes the given range from the key-value store.\nA delete request increments the revision of the key-value store\nand generates a delete event in the event history for every deleted key.",
        "operationId": "DeleteRange",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/etcdserverpbDeleteRangeRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/etcdserverpbDeleteRangeResponse"
            }
          },
          "default": {
            "description": "An unexpected error response",
            "schema": {
              "$ref": "#/definitions/runtimeError"
            }
          }
        }
      }
    },
...
    "/v3/lease/grant": {
      "post": {
        "tags": [
          "Lease"
        ],
        "summary": "LeaseGrant creates a lease which expires if the server does not receive a keepAlive\nwithin a given time to live period. All keys attached to the lease will be expired and\ndeleted if the lease expires. Each expired key generates a delete event in the event history.",
        "operationId": "LeaseGrant",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/etcdserverpbLeaseGrantRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/etcdserverpbLeaseGrantResponse"
            }
          },
          "default": {
            "description": "An unexpected error response",
            "schema": {
              "$ref": "#/definitions/runtimeError"
            }
          }
        }
      }
    },
...
    "/v3/lease/revoke": {
      "post": {
        "tags": [
          "Lease"
        ],
        "summary": "LeaseRevoke revokes a lease. All keys attached to the lease will expire and be deleted.",
        "operationId": "LeaseRevoke",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/etcdserverpbLeaseRevokeRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/etcdserverpbLeaseRevokeResponse"
            }
          },
          "default": {
            "description": "An unexpected error response",
            "schema": {
              "$ref": "#/definitions/runtimeError"
            }
          }
        }
      }
    },
...
    "/v3/maintenance/snapshot": {
      "post": {
        "tags": [
          "Maintenance"
        ],
        "summary": "Snapshot sends a snapshot of the entire backend from a member over a stream to a client.",
        "operationId": "Snapshot",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/etcdserverpbSnapshotRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A successful response.(streaming responses)",
            "schema": {
              "type": "object",
              "title": "Stream result of etcdserverpbSnapshotResponse",
              "properties": {
                "error": {
                  "$ref": "#/definitions/runtimeStreamError"
                },
                "result": {
                  "$ref": "#/definitions/etcdserverpbSnapshotResponse"
                }
              }
            }
          },
          "default": {
            "description": "An unexpected error response",
            "schema": {
              "$ref": "#/definitions/runtimeError"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "etcdserverpbCompactionRequest": {
      "description": "CompactionRequest compacts the key-value store up to a given revision. All superseded keys\nwith a revision less than the compaction revision will be removed.",
      "type": "object",
      "properties": {
        "physical": {
          "description": "physical is set so the RPC will wait until the compaction is physically\napplied to the local database such that compacted entries are totally\nremoved from the backend database.",
          "type": "boolean",
          "format": "boolean"
        },
        "revision": {
          "description": "revision is the key-value store revision for the compaction operation.",
          "type": "string",
          "format": "int64"
        }
      }
    },
    "etcdserverpbCompactionResponse": {
      "type": "object",
      "properties": {
        "header": {
          "$ref": "#/definitions/etcdserverpbResponseHeader"
        }
      }
    },
...
    "etcdserverpbDeleteRangeRequest": {
      "type": "object",
      "properties": {
        "key": {
          "description": "key is the first key to delete in the range.",
          "type": "string",
          "format": "byte"
        },
        "prev_kv": {
          "description": "If prev_kv is set, etcd gets the previous key-value pairs before deleting it.\nThe previous key-value pairs will be returned in the delete response.",
          "type": "boolean",
          "format": "boolean"
        },
        "range_end": {
          "description": "range_end is the key following the last key to delete for the range [key, range_end).\nIf range_end is not given, the range is defined to contain only the key argument.\nIf range_end is one bit larger than the given key, then the range is all the keys\nwith the prefix (the given key).\nIf range_end is '\\0', the range is all keys greater than or equal to the key argument.",
          "type": "string",
          "format": "byte"
        }
      }
    },
    "etcdserverpbDeleteRangeResponse": {
      "type": "object",
      "properties": {
        "deleted": {
          "description": "deleted is the number of keys deleted by the delete range request.",
          "type": "string",
          "format": "int64"
        },
        "header": {
          "$ref": "#/definitions/etcdserverpbResponseHeader"
        },
        "prev_kvs": {
          "description": "if prev_kv is set in the request, the previous key-value pairs will be returned.",
          "type": "array",
          "items": {
            "$ref": "#/definitions/mvccpbKeyValue"
          }
        }
      }
    },
...
    "etcdserverpbLeaseGrantRequest": {
      "type": "object",
      "properties": {
        "ID": {
          "description": "ID is the requested ID for the lease. If ID is set to 0, the lessor chooses an ID.",
          "type": "string",
          "format": "int64"
        },
        "TTL": {
          "description": "TTL is the advisory time-to-live in seconds. Expired lease will return -1.",
          "type": "string",
          "format": "int64"
        }
      }
    },
    "etcdserverpbLeaseGrantResponse": {
      "type": "object",
      "properties": {
        "ID": {
          "description": "ID is the lease ID for the granted lease.",
          "type": "string",
          "format": "int64"
        },
        "TTL": {
          "description": "TTL is the server chosen lease time-to-live in seconds.",
          "type": "string",
          "format": "int64"
        },
        "error": {
          "type": "string"
        },
        "header": {
          "$ref": "#/definitions/etcdserverpbResponseHeader"
        }
      }
    },
...
    "etcdserverpbLeaseRevokeRequest": {
      "type": "object",
      "properties": {
        "ID": {
          "description": "ID is the lease ID to revoke. When the ID is revoked, all associated keys will be deleted.",
          "type": "string",
          "format": "int64"
        }
      }
    },
    "etcdserverpbLeaseRevokeResponse": {
      "type": "object",
      "properties": {
        "header": {
          "$ref": "#/definitions/etcdserverpbResponseHeader"
        }
      }
    },
...
    "etcdserverpbResponseHeader": {
      "type": "object",
      "properties": {
        "cluster_id": {
          "description": "cluster_id is the ID of the cluster which sent the response.",
          "type": "string",
          "format": "uint64"
        },
        "member_id": {
          "description": "member_id is the ID of the member which sent the response.",
          "type": "string",
          "format": "uint64"
        },
        "raft_term": {
          "description": "raft_term is the raft term when the request was applied.",
          "type": "string",
          "format": "uint64"
        },
        "revision": {
          "description": "revision is the key-value store revision when the request was applied.\nFor watch progress responses, the header.revision indicates progress. All future events\nrecieved in this stream are guaranteed to have a higher revision number than the\nheader.revision number.",
          "type": "string",
          "format": "int64"
        }
      }
    },
...
    "etcdserverpbSnapshotRequest": {
      "type": "object"
    },
    "etcdserverpbSnapshotResponse": {
      "type": "object",
      "properties": {
        "blob": {
          "description": "blob contains the next chunk of the snapshot in the snapshot stream.",
          "type": "string",
          "format": "byte"
        },
        "header": {
          "description": "header has the current key-value store information. The first header in the snapshot\nstream indicates the point in time of the snapshot.",
          "$ref": "#/definitions/etcdserverpbResponseHeader"
        },
        "remaining_bytes": {
          "type": "string",
          "format": "uint64",
          "title": "remaining_bytes is the number of blob bytes to be sent after this message"
        }
      }
    },
...
    "mvccpbKeyValue": {
      "type": "object",
      "properties": {
        "create_revision": {
          "description": "create_revision is the revision of last creation on this key.",
          "type": "string",
          "format": "int64"
        },
        "key": {
          "description": "key is the key in bytes. An empty key is not allowed.",
          "type": "string",
          "format": "byte"
        },
        "lease": {
          "description": "lease is the ID of the lease that attached to key.\nWhen the attached lease expires, the key will be deleted.\nIf lease is 0, then no lease is attached to the key.",
          "type": "string",
          "format": "int64"
        },
        "mod_revision": {
          "description": "mod_revision is the revision of last modification on this key.",
          "type": "string",
          "format": "int64"
        },
        "value": {
          "description": "value is the value held by the key, in bytes.",
          "type": "string",
          "format": "byte"
        },
        "version": {
          "description": "version is the version of the key. A deletion resets\nthe version to zero and any modification of the key\nincreases its version.",
          "type": "string",
          "format": "int64"
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "type_url": {
          "type": "string"
        },
        "value": {
          "type": "string",
          "format": "byte"
        }
      }
    },
    "runtimeError": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "details": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/protobufAny"
          }
        },
        "error": {
          "type": "string"
        },
        "message": {
          "type": "string"
        }
      }
    },
    "runtimeStreamError": {
      "type": "object",
      "properties": {
        "details": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/protobufAny"
          }
        },
        "grpc_code": {
          "type": "integer",
          "format": "int32"
        },
        "http_code": {
          "type": "integer",
          "format": "int32"
        },
        "http_status": {
          "type": "string"
        },
        "message": {
          "type": "string"
        }
      }
    }
  },
  "securityDefinitions": {
    "ApiKey": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "security": [
    {
      "ApiKey": []
    }
  ]
}